добавлен полис

+++ src/app/Http/Controllers/RoleController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\http\Response;
+use Illuminate\Support\Facades\Cache;
 
 use App\Models\Role;
 class RoleController extends Controller
@@ -15,7 +16,13 @@ class RoleController extends Controller
      */
     public function index()
     {
+        if (Cache::has('roles')) {
+            // Если данные найдены в кеше, возвращаем их
+            return response()->json(Cache::get('roles'));
+        }
+        $this->authorize('view', Role::class);
         $all_roles = Role::all();
+        Cache::put('roles', $all_roles, now()->addMinutes(10));
         return response()->json($all_roles);
     }
 
@@ -27,6 +34,7 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Role::class);
         $role = Role::create($request->all());
         if (!$role) {
             return response()->json(['error' => 'Role not created'], 422);
@@ -42,6 +50,7 @@ class RoleController extends Controller
      */
     public function show($id)
     {
+        $this->authorize('view', Role::class);
         $role = Role::find($id);
         if (!$role) {
             return response()->json(['error' => 'Role not found'], 404);
@@ -58,6 +67,7 @@ class RoleController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->authorize('update', Role::class);
         $role = Role::find($id);
         if (!$role) {
             return response()->json(['error' => 'Role not found'], 404);
@@ -74,11 +84,12 @@ class RoleController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete', Role::class);
         $role = Role::find($id);
         if (!$role) {
             return response()->json(['error' => 'Role not found'], 404);
         }
-        $copy = $role->toArray();
+        // $copy = $role->toArray();
         $role->delete();
         return response()->json(null, 204);
     }
